+++ app/projektanfragen/page.tsx
@@ -18,7 +18,8 @@ import FormButton from "@/components/form-button";
 import toast from "react-hot-toast";
 import { useAction } from "next-safe-action/hooks";
 import ToastMessage from "@/components/toast-message";
-import { FormSchema, TFormSchema, submitSafeInquiry } from "./_lib/actions";
+import { FormSchema, TFormSchema } from "./_lib/validation";
+import { submitSafeInquiry } from "./_lib/actions";
 import { FileState, MultiFileDropzone } from "@/components/multi-file-dropzone";
 import { EdgeStoreApiClientError } from "@edgestore/react/shared";
 import { formatFileSize } from "@edgestore/react/utils";
