+++ lib/safe-action-clients.ts
@@ -1,4 +1,4 @@
 import { createSafeActionClient } from "next-safe-action";
 
 // This is our base client.
-export const action = createSafeActionClient();
+export const actionClient = createSafeActionClient();
